Move Nutrition AI to a dedicated page and add reusable NutritionAI component (#424)
